Not much has changed. But trying to streamline computation of global poses. Added class GOP but still under-construction. Need to get this done soon.
Completed through analysis of pose and that drift. Turns out it was
all about the units and scale. The translation units were mm (usually a
few 1000s) and rotations are expressed in radians (usually less than
0.1). This is causing numerical problems in the sense the gradient
descent cannot get to the correct translation. This was something I had
actually observed before when the translation was actually not working
well. (Although I was silent about it, which is not turning up as a
problem).

The way forward is
1) To use translation in meters to make it to better scale. This can be
easily achieved by scaling the Z when we compute the 3d co-ordinates.
Thankfully we evaluate the 3d co-ordinates in enlistRefEdgePts() no
where else. No need to worry about the reprojections.
2) In the gradient method, the trust region might be causing trouble.
Possibly make different thresholds for rotation part and translation
part. Think about the math before proceeding in this direction.