Tunning for the fr2/desk sequence. Now the trajectory has a similar shape to the ground truth except for the a different plane :/
Following things done in this version :
-1- Convert everything into meter for numerical stability and to avoid
the scaling problem caused due to totally different scales of numbers
for rotation and translation
-2- Getting rid of the lambda & LM heuristic on step size. Instead
introduced "square summable but not summable sequence (1/k)" of step size
as described in the Boyd's document.
-3- Fixed the floating point error comming because of the LM heuristic.
Although that part is commented out. Ideally I would like to have a
-3- Some other minor fixes

TODO
-1- The change of keyframe program flow is flawed. At the moment, we set
now frame as ref frame if the estimates seem to be bad. The bad part is
we in-coperate these bad estimates into the global trajectory. Instead,
we should be throwing away this bad estimate and setting previously
choosen now frame (1 step before this bad estimation) as the new
reference frame. Suggested by Shen.
-2- Test on more sequences.