Implemented interpolation of distance transform. Can be disabled with a preprocessor define.
TODO
I am now suspecting that precision is not enough for rotation matrix and
to some extent the translation vector. More over the problem gets worse
as the cR,cT form a feedback loop over hunders of iterations.

Following possible solutions
a. rotationize the cR at every iteration and/or only the best cR.
b. templatify the GOP class to handle pose in double precision.